fully commit to using server actions
