--- app/api/helpers.ts
@@ -1,5 +1,0 @@
-const API_URL = process.env.NEXT_PUBLIC_VERCEL_URL
-  ? `https://${process.env.NEXT_PUBLIC_VERCEL_URL}`
-  : 'http://localhost:3000';
-
-export const getApiUrl = (path: string) => `${API_URL}/api${path}`;
